refactor: DebugTimberTree util로 패키지 이동
