+++ app/src/main/java/com/pomonyang/mohanyang/MohaNyangApplication.kt
@@ -1,7 +1,7 @@
 package com.pomonyang.mohanyang
 
 import android.app.Application
-import com.pomonyang.mohanyang.ui.DebugTimberTree
+import com.pomonyang.mohanyang.util.DebugTimberTree
 import dagger.hilt.android.HiltAndroidApp
 import timber.log.Timber
